Rename HeroNavMenu to NavigateHeroes
Updated import paths and renamed files to reflect the change from HeroNavMenu to NavigateHeroes. This improves clarity and better represents the functionality provided by the module.
