+++ frontend-next-migration/src/features/NavigateHeroes/ui/HeroNavMenu.tsx
